pyproject.toml
[tool.poetry]
name = "lifeshare-back"
version = "0.1.0"
description = ""
authors = ["Yahialabeeb <yahia.labeeb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
django-cors-headers = "^3.10.1"
djangorestframework-simplejwt = "^5.0.0"
django-phone-field = "^1.8.1"
Pillow = "^8.4.0"
django-rest-passwordreset = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
